Checkbox group tests
